addresses BC and segment length issues in minsnap
minsnap boundary conditions (in jerk etc) were improperly constrained and
was identified to fail on segment lengths that were too short. This addresses
those and adds and auto trajectory selection implementation (generate_trajectory_auto.m)
to address these concerns. Propagates these changes to test and simulation entry points

Along the way, a sensitivity to the figure eight trajectory was identified--it seems the IC
for this trajectory results in some yaw discontinuty which causes the simulation
to blow up. This introduced a simpler square trajectory and uses that in the
inttest_trajectory_closedloop.m demonstrator to allow some meaningful comparison
of the trajectory solution implementations